split metered, add Sink impl
